% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis.R
\name{clusterAnalysis}
\alias{clusterAnalysis}
\title{clusterAnalysis}
\usage{
clusterAnalysis(expres, genes, samples, sampleNames, comparisonName,
  anotPackage, my.symbols = NULL, outputDir, fileOfLinks, numClusters,
  rowDistance, colDistance, RowVals = TRUE, ColVals = TRUE, colorsSet,
  colsForGroups, escala, densityInfo = "none", cexForColumns, cexForRows,
  plotProfiles = FALSE, Title = "", csvType = NULL)
}
\arguments{
\item{expres}{Expresion oused to perform the clusters.}

\item{genes}{}

\item{samples}{}

\item{sampleNames}{Name of the samples of the study.}

\item{comparisonName}{Name of the comparison. To identify the output}

\item{anotPackage}{Annotation package.}

\item{outputDir}{Path of the file created.}

\item{fileOfLinks}{Name of the links file.}

\item{rowDistance}{Row distance in the clusters.}

\item{colDistance}{Column distance in the clusters.}

\item{colsForGroups}{Colors of each group for the plots.}

\item{escala}{}

\item{cexForColumns}{Cex for the columns.}

\item{cexForRows}{Cex for the rows.}

\item{plotProfiles}{}

\item{Title}{Title of the plot.}

\item{csvType}{Csv type of the output.}
}
\value{
hm (?¿?¿?¿?)
}
\description{
Function to make the cluster analysis.
}
