% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doClusterAnalysis.R
\name{doClusterAnalysis}
\alias{doClusterAnalysis}
\title{doClusterAnalysis}
\usage{
doClusterAnalysis(clustPar)
}
\arguments{
\item{clustPar}{List object that contains the parameters.}
}
\description{
Function to perform cluster analysis
}
\examples{
for(ix in 1:length(clustParsList))
{hm.Estudi <- doClusterAnalysis(clustParsList[ix])}
}
