% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGeneAnnotation.R
\name{doGeneAnnotation}
\alias{doGeneAnnotation}
\title{doGeneAnnotation}
\usage{
doGeneAnnotation(AnotList)
}
\arguments{
\item{anotList}{List object that contains the parameters}
}
\value{
genesAnnotated (??)
}
\description{
Function that annotates a data set.
}
