% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGOAnalysis.R
\name{doGOAnalysis}
\alias{doGOAnalysis}
\title{doGOAnalysis}
\usage{
doGOAnalysis(GOPar)
}
\arguments{
\item{GOParsList}{List object that contains the parameters needed to carry out the analysis.}
}
\value{
GOResult
}
\description{
Function to interate with GOAnalysis
}
