% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doKEGGAnalysis.R
\name{doKEGGAnalysis}
\alias{doKEGGAnalysis}
\title{doKEGGAnalysis}
\usage{
doKEGGAnalysis(KEGGPar)
}
\arguments{
\item{KEGGPar}{List object that contains the parameters.}
}
\value{
KEGGResult
}
\description{
Function to interate with KEGGAnalysis
}
\examples{
 for(i in 1:length(KEGGParsList))
 {
 KEGGList <- doKEGGAnalysis(KEGGParsList[i])
 }
}
