% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doLmAnalysis.R, R/lmAnalysis.R
\name{doLmAnalysis}
\alias{doLmAnalysis}
\alias{doLmAnalysis}
\title{doLmAnalysis}
\usage{
doLmAnalysis(lmPar)

doLmAnalysis(lmPar)
}
\arguments{
\item{lmPar}{list object that contains the parameters}

\item{lmPar}{list object that contains the parameters}
}
\description{
doLmAnalysis

doLmAnalysis
}
