% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMultCompAnalysis.R
\name{doMultCompAnalysis}
\alias{doMultCompAnalysis}
\title{doMultCompAnalysis}
\usage{
doMultCompAnalysis(mcPar)
}
\arguments{
\item{mcPar}{List object that contains the parameters.}
}
\value{
genelist
}
\description{
Function for make the comparative analysis.
}
