% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnnotation.R
\name{GeneAnnotation}
\alias{GeneAnnotation}
\title{GeneAnnotation}
\usage{
GeneAnnotation(egIDs, anotPackage, toHTML = TRUE, outputDir,
  filename = "annotations", myTitle = "Annotations for all genes analyzed",
  specie = "Homo_sapiens", info2show = c("Affymetrix", "EntrezGene"),
  linksFile, maxGenes = NULL, ...)
}
\arguments{
\item{anotPackage}{Annotation package.}

\item{toHTML}{If TRUE a html file will be created.}

\item{outputDir}{Path of the file created.}

\item{filename}{Name of the file}

\item{myTitle}{Title of ???}

\item{specie}{Specie}

\item{linksFile}{Name of the LinksFile}
}
\value{
It returns the time that the process least.
}
\description{
Function to annotate genes
}
