% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoAnalysis.R
\name{GOAnalysis}
\alias{GOAnalysis}
\title{GOAnalysis}
\usage{
GOAnalysis(fitMain, whichContrasts, comparison.Name, outputDir, anotPackage,
  my.IDs, addGeneNames = TRUE, fileOfLinks, thrLogFC = NULL,
  cutoffMethod = c("adjusted", "unadjusted"), P.Value.cutoff = rep(0.05,
  length(whichContrasts)), pval = 0.05, min.count = 3,
  ontologias = c("MF", "BP", "CC"), testDirections = c("over", "under"),
  minNumGens = 0)
}
\arguments{
\item{comparison.Name}{Name of the comparison. To identify the output.}

\item{outputDir}{Path of the file created.}

\item{anotPackage}{Annotation package.}

\item{addGeneNames}{Vector of the gene names.}

\item{fileOfLinks}{Name of the links file.}

\item{thrLogFC}{Threshold for the logarithm Fold Change}

\item{P.Value.cutoff}{Maximum value for the p.value.}

\item{minNumGens}{Mimimum number of genes.}
}
\description{
Function that do GOAnalysis
}
\examples{
 
\dontrun{
library(GOstats)
GOResult<-BasicP::GOAnalysis(fitMain = fitMain, whichContrasts = wCont, 
comparison.Name = "Estudi", outputDir = outputDir, anotPackage = "org.Hs.eg", 
my.IDs = entrezTable, addGeneNames = TRUE, fileOfLinks = linksFile, thrLogFC = 1, 
cutoffMethod = "adjusted", P.Value.cutoff = rep(0.05, length(wCont)), pval = 0.01, 
min.count = 3, ontologias = c("MF", "BP", "CC"), testDirections = c("over", "under"), 
minNumGens = 0)
}
}
