% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGAnalysis.R
\name{KEGGAnalysis}
\alias{KEGGAnalysis}
\title{KEGGAnalysis}
\usage{
KEGGAnalysis(fitMain, whichContrasts, comparison.Name, outputDir, anotPackage,
  organisme, my.IDs, addGeneNames = TRUE, fileOfLinks,
  cutoffMethod = c("adjusted", "unadjusted"), P.Value.cutoff = rep(0.05,
  length(whichContrasts)), pval = 0.05, thrLogFC = NULL, minNumGens = 0)
}
\description{
Function to make KEGG analysis.
}
