% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmAnalysis.R
\name{lmAnalysis}
\alias{lmAnalysis}
\title{lmAnalysis}
\usage{
lmAnalysis(exprs.filtered, design, cont.matrix,
  contrasts2test = 1:ncol(cont.matrix), anotPackage,
  Expressions_And_Top = TRUE, showParams = FALSE, use.dupCorr = FALSE,
  block = NULL, nDups = 1, comparison = "", outputDir = ".",
  ENTREZIDs = NULL, SYMBOLIDs = NULL, linksFile, fitFileName, csvType,
  rows2HTML = NULL, anotFileName)
}
\arguments{
\item{exprs.filtered}{Dataset of filtered and normalized data.}

\item{design}{Design matrix.}

\item{cont.matrix}{Contrast matrix.}

\item{contrasts2test}{Numeric vector that contains the number of the columns for the contrasts to analyze.}

\item{anotPackage}{Annotation package.}

\item{Expressions_And_Top}{If is TRUE it will do "Expresions" and the toptable.}

\item{showParams}{If FALSe it won't show the lmFit parameters in the ExpressionsAndToptable.}

\item{use.dupCorr}{???}

\item{block}{???}

\item{nDups}{???}

\item{comparison}{Name of the comparations in all contrasts.}

\item{outputDir}{Path of the file created.}

\item{ENTREZIDs}{Name of the file for Entrez genes.}

\item{SYMBOLIDs}{Name of the file for gene Symbols .}

\item{linksFile}{Name of the file linksFile.}

\item{fitFileName}{???}

\item{csvType}{type of csv to store the data.}

\item{rows2HTML}{???}

\item{anotFileName}{???}
}
\description{
Function to make the linear model analysis.
}
