% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleComp.R
\name{multipleComp}
\alias{multipleComp}
\title{multipleComp}
\usage{
multipleComp(fitMain, whichContrasts, comparisonName, titleText, outputDir,
  anotPackage, my.symbols = NULL, linksFile, multCompMethod = "separate",
  adjustMethod = "none", selectionType = c("any", "all", "anyDown", "anyUp",
  "allDown", "allUp"), P.Value.cutoff = 0.05, plotVenn, colsVenn = NULL,
  vennColors, cexVenn = 1, geneListFName = paste("geneList", comparisonName,
  "Rda", sep = "."), csvType = NULL, minLFC = 0)
}
\arguments{
\item{whichContrasts}{Vector that determie the contrast to do.}

\item{comparisonName}{Name of the comparisons.}

\item{outputDir}{Path of the file created.}

\item{anotPackage}{Annotation package.}

\item{linksFile}{Name of the LinksFile.}

\item{multCompMethod}{Method to make the comparation.}

\item{adjustMethod}{Method to apply in the adjust.}

\item{P.Value.cutoff}{Cutoff value for p.value.}
}
\value{
geneList (?¿?)
}
\description{
Function to make multiple comparations.
}
