% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doLmAnalysis.R, R/lmAnalysis.R
\name{doLmAnalysis}
\alias{doLmAnalysis}
\alias{doLmAnalysis}
\title{doLmAnalysis}
\usage{
doLmAnalysis(lmPar)

doLmAnalysis(lmPar)
}
\arguments{
\item{lmPar}{list object that contains the parameters needed to carry out the analysis.}

\item{lmPar}{list object that contains the parameters}
}
\description{
Function to make the Linear model analysis. The unique parameter is lmPar list.

doLmAnalysis
}
