% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleComp.R
\name{multipleComp}
\alias{multipleComp}
\title{multipleComp}
\usage{
multipleComp(fitMain, whichContrasts, comparisonName, titleText, outputDir,
  anotPackage, my.symbols = NULL, linksFile, multCompMethod = "separate",
  adjustMethod = "none", selectionType = c("any", "all", "anyDown", "anyUp",
  "allDown", "allUp"), P.Value.cutoff = 0.05, plotVenn, colsVenn = NULL,
  vennColors, cexVenn = 1, geneListFName = paste("geneList", comparisonName,
  "Rda", sep = "."), csvType = NULL, minLFC = 0)
}
\arguments{
\item{whichContrasts}{Vector that determie the contrast to do.}

\item{comparisonName}{Name of the comparisons.}

\item{outputDir}{Path of the file created.}

\item{anotPackage}{Annotation package.}

\item{linksFile}{Name of the LinksFile.}

\item{multCompMethod}{Method to make the comparation.}

\item{adjustMethod}{Method to apply in the adjust.}

\item{P.Value.cutoff}{Cutoff value for p.value.}
}
\description{
Function to make multiple comparations.
}
\examples{
\dontrun{
compName <- c("Group1")
wCont <- 1:3
pValCutOff <- c(0.01)
adjMethod <- c("none")
minLogFoldChange <- c(1)
load("./ResultsDir/Symbols.Rda")
titleText = paste("for", ifelse(adjMethod=="none","p-values","adj. p-values"), "<", pValCutOff, "and |logFC| >", minLogFoldChange, sep = " ")
geneListFName = paste("geneList",compName,ifelse(adjMethod=="none","pvalues","adj-pvalues"),"LT",pValCutOff,"Rda",sep = ".")

geneList <-  BasicP::multipleComp(fitMain = fitMain, whichContrasts = wCont, comparisonName = compName, titleText = titleText,
outputDir = outputDir,  anotPackage = "org.Hs.eg", my.symbols = symbolsTable, linksFile = linksFile, multCompMethod = "separate",
adjustMethod = adjMethod, selectionType = "any", P.Value.cutoff = pValCutOff, plotVenn = TRUE, colsVenn = NULL, 
vennColors = c("red","yellow","green","blue","pink") , cexVenn = 1, geneListFName=geneListFName, csvType=csvType, minLFC=minLogFoldChange)
}
}
